+++ src/crawlers/basic_crawler.js
@@ -635,6 +635,21 @@ export class BasicCrawler {
     }
 
     /**
+     * @param {Array<Hook>} hooks
+     * @param  {*} args
+     * @ignore
+     * @protected
+     * @internal
+     */
+    async _executeHooks(hooks, ...args) {
+        if (Array.isArray(hooks) && hooks.length) {
+            for (const hook of hooks) {
+                await hook(...args);
+            }
+        }
+    }
+
+    /**
      * Function for cleaning up after all request are processed.
      * @ignore
      */