+++ src/utils_request.js
@@ -49,6 +49,8 @@ import { IncomingMessage } from 'http';
  *  expect HTTP2 connections, because browsers use HTTP2 by default. It will automatically downgrade
  *  to HTTP/1.1 for websites that do not support HTTP2. For Node 10 this option is always set to `false`
  *  because Node 10 does not support HTTP2 very well. Upgrade to Node 12 for better performance.
+ * @property {boolean} [forceUrlEncoding]
+ *   Automatically encode URLs via `encodeURI()` before resolving them.
  */
 
 /**
@@ -119,6 +121,7 @@ export const requestAsBrowser = async (options = {}) => {
         headerGeneratorOptions: ow.optional.object,
         stream: ow.optional.boolean,
         decodeBody: ow.optional.boolean,
+        forceUrlEncoding: ow.optional.boolean,
     }));
 
     ow(options, 'RequestAsBrowserOptions', ow.object.validate((opts) => ({
@@ -146,6 +149,7 @@ export const requestAsBrowser = async (options = {}) => {
         throwOnHttpErrors = false,
         stream = false,
         decodeBody = true,
+        forceUrlEncoding,
         ...gotParams
     } = options;
 
@@ -180,7 +184,13 @@ export const requestAsBrowser = async (options = {}) => {
         };
     }
 
-    if (!gotScrapingOptions.isStream) return gotScraping(gotScrapingOptions);
+    if (forceUrlEncoding) {
+        gotScrapingOptions.url = encodeURI(gotScrapingOptions.url);
+    }
+
+    if (!gotScrapingOptions.isStream) {
+        return gotScraping(gotScrapingOptions);
+    }
 
     // abortFunction must be handled separately for streams :(
     const duplexStream = await gotScraping(gotScrapingOptions);